LayoutTests/http/tests/media/fairplay/legacy-fairplay-error-expected.txt


EVENT(webkitneedkey)
Create media keys:
Attempt to create key session with invalid certificate:
EVENT(webkitkeyerror)
END OF TEST
